Some entities and basic logic added

+++ source/CoffeeBank/Coffee.Entities/CreditRequest.cs
@@ -2,10 +2,12 @@
 {
     public class CreditRequest
     {
-        public long Sum { get; set; } //In belorussian rubles only
+        public decimal Sum { get; set; } //In belorussian rubles only
 
         public PassportInfo PassportInfo { get; set; }
 
         public SalaryInfo SalaryInfo { get; set; }
+
+        //...some other info
     }
 }
